Create README.md
Duckie.ai project
Create a .NET MAUI Blazer Hybrid Application where user create a profile and then create an intuitive and interactive family tree or work leadership tree. A user can a person(s) by adding a name, uploading an image from local storage, set the person's gender, date of birth, (and end date if deceased), and a one to two paragraph biography. These attributes about the person must be grouped into one object and the object must be editable and easily connectable to other object(s) as desired by the user using lines as node. the image or the person must be presented a cirlce that shows their head and shoulders. the user can create multiple family trees and must be able to save them, export them and share them via email. the person objects must be interactive where the first three words of the biography are visible but the rest can be revealed by clicking the biography title or person's image. Use the new features of .NET 8 Preview framework for a modern look and feel to the multiplatform application.